[Refactoring] Changed station namespace
