+++ Assets/Features/StationLogic/UI/OpenStationButton.cs
@@ -2,7 +2,7 @@
 using UnityEngine;
 using UnityEngine.UI;
 
-namespace Features.Station.UI
+namespace Features.StationLogic.UI
 {
     public class OpenStationButton : MonoBehaviour
     {
