+++ Assets/Features/StationLogic/UI/RemainingEnqueuesCounter.cs
@@ -2,7 +2,7 @@
 using UnityEngine;
 using UnityEngine.UI;
 
-namespace Features.Station.UI
+namespace Features.StationLogic.UI
 {
     public class RemainingEnqueuesCounter : MonoBehaviour
     {
