+++ Assets/Features/StationLogic/UI/ResetStationButton.cs
@@ -1,7 +1,7 @@
 ﻿using UnityEngine;
 using UnityEngine.UI;
 
-namespace Features.Station.UI
+namespace Features.StationLogic.UI
 {
     public class ResetStationButton : MonoBehaviour
     {
